+++ test/index.spec.js
@@ -45,6 +45,22 @@ let cases = [
     `,
   ],
   [
+    'optional chaining',
+    `
+    let obj = {
+      foo: {
+        bar: {
+          baz: 42,
+        },
+      },
+    };
+
+    let baz = obj?.foo?.bar?.baz; // 42
+
+    let safe = obj?.qux?.baz; // undefined
+    `,
+  ],
+  [
     'Runtime ponyfills',
     `
       new Map()